refactor: unified transform component across engine
Replaces direct Transform component access with GetRootComponent and standardizes use of world position, rotation, and scale throughout engine components. Object scale and rotation are now managed at the GameObject level, and rendering, collision, and sprite logic have been updated to use these properties. This improves consistency, simplifies component interfaces, and prepares the codebase for more robust transform hierarchies.

+++ Engine/Engine/Input/Controller.h
@@ -6,7 +6,7 @@ namespace diji
     class Controller final
     {
     public:
-        enum class Button : unsigned int
+        enum class Button : uint16_t
         {
             DPadUp = 0x0001,
             DPadDown = 0x0002,
